Inital Test Refactor
