+++ custom_components/husqvarna_automower/tests/test_device_tracker.py
@@ -20,6 +20,8 @@ from .const import (
 )
 from .test_init import configure_application_credentials
 
+from .test_common import setup_entity
+
 
 @pytest.mark.asyncio
 async def setup_device_tracker(hass: HomeAssistant, empty_positions: bool = False):
@@ -69,8 +71,15 @@ async def test_device_tracker_setup_no_pos(hass: HomeAssistant):
     """test device tracker without positions."""
 
     # Without positions
-    await setup_device_tracker(hass, empty_positions=True)
-    assert len(hass.config_entries.async_entries(DOMAIN)) == 1
+    session = deepcopy(AUTOMOWER_SM_SESSION_DATA)
+    session["data"][MWR_ONE_IDX]["attributes"]["positions"] = []
+
+    with patch(
+        "custom_components.husqvarna_automower.tests.test_common.AUTOMOWER_SM_SESSION_DATA",
+        session,
+    ):
+        await setup_entity(hass)
+        assert len(hass.config_entries.async_entries(DOMAIN)) == 1
 
     coordinator = hass.data[DOMAIN]["automower_test"]
     device_tracker = AutomowerTracker(coordinator, MWR_ONE_IDX)
@@ -79,7 +88,7 @@ async def test_device_tracker_setup_no_pos(hass: HomeAssistant):
 @pytest.mark.asyncio
 async def test_device_tracker_pos(hass: HomeAssistant):
     """test device tracker with positions."""
-    await setup_device_tracker(hass, empty_positions=False)
+    await setup_entity(hass)
     coordinator = hass.data[DOMAIN]["automower_test"]
     device_tracker = AutomowerTracker(coordinator, MWR_ONE_IDX)
     assert device_tracker._attr_unique_id == f"{MWR_ONE_ID}_dt"
